+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 - Bug #746: Refactor `DMLQueryBuilder` and fix: unique indexes in `upsert()`, column names with table name and brackets, `batchInsert()` with associative arrays, enhanced documentation of `batchInsert()` and `update()` (@Tigrov)
 - Bug #751: Fix collected debug actions (@xepozz)
+- Chg #755: Deprecate `TableSchemaInterface::compositeForeignKey()` (@Tigrov) 
+- Enh #756: Refactor `Quoter` (@Tigrov) 
+- Bug #756: Fix `Quoter::quoteSql()` for SQL containing table with prefix (@Tigrov)
+- Bug #756: Fix `Quoter::getTableNameParts()` for cases when different quotes for tables and columns (@Tigrov)
+- Bug #756: Fix `Quoter::quoteTableName()` for sub-query with alias (@Tigrov)
+- Bug #761: Quote aliases of CTE in `WITH` queries (@Tigrov)
+- Chg #765: Deprecate `SchemaInterface::TYPE_JSONB` (@Tigrov)
 
 ## 1.1.1 August 16, 2023